Aligns w/ 15.5 MOOS-IvP new Geodesy structure as a wrapping of proj4. We're using proj4.9.2 here to be able to use cmake w/ it.
