+++ CMakeLists.txt
@@ -81,7 +81,10 @@ if (PROJECT_EXPORT_LIBS)
     foreach(EXPORT_LIB ${PROJECT_EXPORT_LIBS})
         file(APPEND ${CONFIG_FILE} "\n# Export variables for ${EXPORT_LIB} Library\n")
         file(APPEND ${CONFIG_FILE} "set(${EXPORT_LIB}_LIBRARIES \"${${EXPORT_LIB}_LIBRARIES}\")\n")
+
         file(APPEND ${CONFIG_FILE} "set(${EXPORT_LIB}_INCLUDE_DIRS \"${${EXPORT_LIB}_INCLUDE_DIRS}\")\n")
+        file(APPEND ${CONFIG_FILE} "set(${EXPORT_LIB}_LIBRARY_PATH \"${${EXPORT_LIB}_DEPEND_INCLUDE_DIRS}/../../lib\")\n")
+
         file(APPEND ${CONFIG_FILE} "set(${EXPORT_LIB}_DEPEND_LIBRARIES \"${${EXPORT_LIB}_DEPEND_LIBRARIES}\")\n")
         file(APPEND ${CONFIG_FILE} "set(${EXPORT_LIB}_DEPEND_INCLUDE_DIRS \"${${EXPORT_LIB}_DEPEND_INCLUDE_DIRS}\")\n")
     endforeach()
