+++ appveyor.yml
@@ -1,0 +1,50 @@
+version: devel-{build}
+
+platform:
+    - x64
+    - x86
+
+environment:
+    matrix:
+    - BUILD_TYPE: cmake
+      VS_VERSION: Visual Studio 11 2012
+    
+    - BUILD_TYPE: cmake
+      VS_VERSION: Visual Studio 12 2013
+      
+    - BUILD_TYPE: cmake
+      VS_VERSION: Visual Studio 14 2015
+
+shallow_clone: true
+
+init:
+    - if "%platform%" == "x64" SET VS_FULL=%VS_VERSION% Win64
+    - if "%platform%" == "x86" SET VS_FULL=%VS_VERSION%
+    - cd ..
+    - git clone -b wOnlineCI --depth=1 https://github.com/msis/core-moos
+    - cd core-moos
+    - mkdir build
+    - cd build
+    - cmake -G "%VS_FULL%" -DENABLE_EXPORT=ON -DUSE_ASYNC_COMMS=ON -DTIME_WARP_AGGLOMERATION=0.4 -DCMAKE_BUILD_TYPE=$BUILD_TYPE ../
+    - cmake --build . --config Release
+    - cmake --build . --config Release --target install
+    - cd %APPVEYOR_BUILD_FOLDER%/..
+    - git clone -b 4.9.2 --depth=1 https://github.com/osgeo/proj.4 proj-4
+    - cd proj-4
+    - mkdir build
+    - cd build 
+    - cmake -DBUILD_TESTING=OFF -DBUILD_PROJ4_TESTS=OFF ../
+    - cmake --build . --config Release
+    - cmake --build . --config Release --target install
+    - set PROJ4_INCLUDE_DIR=%PWD%/../src
+    - set PROJ4_LIB_DIR=%PWD%/lib
+    - cd %APPVEYOR_BUILD_FOLDER%
+    - mkdir build
+
+build_script:
+    - cd build
+    - cmake -G "%VS_FULL%" -DPROJ4_INCLUDE_DIRS=$PROJ4_INCLUDE_DIR -DPROJ4_LIB_PATH=$PROJ4_LIB_DIR -DCMAKE_BUILD_TYPE=$BUILD_TYPE ../
+    - cmake --build . --config Release
+
+on_success:
+    - cmake --build . --config Release --target install
